Define sum of multiples as sum over list comprehension
Our multiples_up_to method is a pretty simple list comprehension. We
call that for each factor, flatten, and sum.

The one trick was realizing that multiplying the factor by the limit
will always meet/exceed the limit, so we can reduce our number of
possible multiples to be <= limit.